[libc] move src/network to src/arpa/inet
So that docgen can find our implementations.

Fixes: #119272
