+++ libc/src/CMakeLists.txt
@@ -33,12 +33,12 @@ if(NOT LLVM_LIBC_FULL_BUILD)
   return()
 endif()
 
+add_subdirectory(arpa)
 add_subdirectory(assert)
 add_subdirectory(compiler)
-add_subdirectory(network)
+add_subdirectory(locale)
 add_subdirectory(search)
 add_subdirectory(setjmp)
 add_subdirectory(signal)
 add_subdirectory(spawn)
 add_subdirectory(threads)
-add_subdirectory(locale)
