+++ libc/src/arpa/inet/ntohs.cpp
@@ -1,4 +1,4 @@
-//===-- Implementation header of htons --------------------------*- C++ -*-===//
+//===-- Implementation of ntohs function ----------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,16 +6,18 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_NETWORK_HTONS_H
-#define LLVM_LIBC_SRC_NETWORK_HTONS_H
-
+#include "src/arpa/inet/ntohs.h"
+#include "src/__support/common.h"
+#include "src/__support/endian_internal.h"
 #include "src/__support/macros/config.h"
-#include <stdint.h>
 
 namespace LIBC_NAMESPACE_DECL {
 
-uint16_t htons(uint16_t hostshort);
+LLVM_LIBC_FUNCTION(uint16_t, ntohs, (uint16_t netshort)) {
+  if constexpr (Endian::IS_LITTLE)
+    return __builtin_bswap16(netshort);
+  else
+    return netshort;
+}
 
 } // namespace LIBC_NAMESPACE_DECL
-
-#endif // LLVM_LIBC_SRC_NETWORK_HTONS_H
