+++ libc/test/src/arpa/inet/CMakeLists.txt
@@ -1,0 +1,53 @@
+add_custom_target(libc_arpa_inet_unittests)
+
+add_libc_unittest(
+  htonl
+  SUITE
+    libc_arpa_inet_unittests
+  SRCS
+    htonl_test.cpp
+  CXX_STANDARD
+    20
+  DEPENDS
+    libc.src.arpa.inet.htonl
+    libc.src.arpa.inet.ntohl
+)
+
+add_libc_unittest(
+  htons
+  SUITE
+    libc_arpa_inet_unittests
+  SRCS
+    htons_test.cpp
+  CXX_STANDARD
+    20
+  DEPENDS
+    libc.src.arpa.inet.htons
+    libc.src.arpa.inet.ntohs
+)
+
+add_libc_unittest(
+  ntohl
+  SUITE
+    libc_arpa_inet_unittests
+  SRCS
+    ntohl_test.cpp
+  CXX_STANDARD
+    20
+  DEPENDS
+    libc.src.arpa.inet.htonl
+    libc.src.arpa.inet.ntohl
+)
+
+add_libc_unittest(
+  ntohs
+  SUITE
+    libc_arpa_inet_unittests
+  SRCS
+    ntohs_test.cpp
+  CXX_STANDARD
+    20
+  DEPENDS
+    libc.src.arpa.inet.htons
+    libc.src.arpa.inet.ntohs
+)
